extract reusable utilities from editor.tsx component
